+++ decomp/ST.exe/functions/00475530/decomp.c
@@ -92,16 +92,16 @@ LAB_00475545:
             sVar18 = 0;
             bVar19 = 0;
             lVar10 = Library::MSVCRT::__ftol();
-            iVar7 = (int)local_20 + (short)this->field_0045 + local_c + (int)(short)lVar10;
-            iVar8 = (int)(short)this->field_0043 - (int)local_24._2_2_;
-            iVar9 = (int)(short)local_24 + iVar4 + (short)this->field_0041;
+            iVar7 = (int)local_20 + this->field_0045 + local_c + (int)(short)lVar10;
+            iVar8 = (int)this->field_0043 - (int)local_24._2_2_;
+            iVar9 = (int)(short)local_24 + iVar4 + this->field_0041;
           }
           else {
-            iVar7 = (int)local_20 + local_c + (short)this->field_0045;
+            iVar7 = (int)local_20 + local_c + this->field_0045;
             local_24._2_2_ = (short)((uint)uVar2 >> 0x10);
-            iVar8 = (int)(short)this->field_0043 - (int)local_24._2_2_;
+            iVar8 = (int)this->field_0043 - (int)local_24._2_2_;
             local_24._0_2_ = (short)uVar2;
-            iVar9 = (int)(short)local_24 + iVar4 + (short)this->field_0041;
+            iVar9 = (int)(short)local_24 + iVar4 + this->field_0041;
           }
           TraksClassTy::TraksCreate
                     (DAT_00802a7c,1,2,7,iVar9,iVar8 + ((uVar5 >> 0x10) % 7 - 3),iVar7,sVar11,sVar12,
